+++ framework/src/main/java/OrderOperations.java
@@ -1,11 +1,11 @@
 public interface OrderOperations<T extends Order> {
     T buyNow();
     T sellNow();
-    T buyStop(int pips) throws TradingLogicException;
-    T sellStop(int pips) throws TradingLogicException;
-    T buyLimit(int pips) throws TradingLogicException;
-    T sellLimit(int pips) throws TradingLogicException;
-    T stopLoss(int pips) throws TradingLogicException;
-    T makeProfit(int pips) throws TradingLogicException;
+    T buyStop(double pips) throws TradingLogicException;
+    T sellStop(double pips) throws TradingLogicException;
+    T buyLimit(double pips) throws TradingLogicException;
+    T sellLimit(double pips) throws TradingLogicException;
+    T stopLoss(double pips) throws TradingLogicException;
+    T makeProfit(double pips) throws TradingLogicException;
     T close();
 }
